MA - 113- This was the parent task for creating firebase, sharing with the team, and starting authetication process.
MA-114 - was the task for if the user is getting and email when creating an account. Now, users get email when creating an account that they need to vertify using link sent to them.

MA-115 - This task was for making sure firebase is being filled with tegh appropirate data and not allowing users to creat accounts using the same email. Also making sure the vertification process happens before users get access to the app. It will not allow users access before making sure the email is not in use by another account.