Ramin Selseleh MA - 116 completed.
 This task was for setting up the forget password
functionality. Now when user clicks forget password, they are sent a link
to reset their passwrod if they have an account with us. If not, they will
get an error message. During this task, I also fixed errors regarding
text coclor, phone number field was not taking in numbers, but rather any string
Now, it will only take in a phine number. Also, the action button in the
login page was not going to the correct page. Now it is going to the right forget
password page.