Update message_page_model.dart
The MessagePageModel class helps manage the data and setup for the MessagePageWidget. It includes basic functions to start (initState) and end (dispose) the model, which can be used to set up or clean up resources if needed. This model is a foundation to add more functions or data in the future to support the MessagePageWidget.