MA-224
Cancel appointment button must be configured to handle the cancelation of recurring appointments. This will be verified by making sure the calls are being sent to the database and they are being removed as requested by the user. Additionally the button locations needs be centered under the confirm appointment button to make the user interface more uniform